Some improvements for LLM/Whisper infer requests
- Move utils functions for infer request to separate file
- Change the way whisper model is identified
- Remove debug dumps
